+++ broker/tcp/src/stream.cc
@@ -57,7 +57,8 @@ stream::stream(const tcp_config::pointer& conf)
                        _conf->get_port());
   log_v2::tcp()->info(
       "{} TCP streams are configured on a thread pool of {} threads",
-      _total_tcp_count, pool::instance().get_pool_size());
+      static_cast<uint32_t>(_total_tcp_count),
+      pool::instance().get_pool_size());
 }
 
 /**
@@ -76,7 +77,8 @@ stream::stream(const tcp_connection::pointer& conn,
                       _conf->get_port());
   log_v2::tcp()->info(
       "{} TCP streams are configured on a thread pool of {} threads",
-      _total_tcp_count, pool::instance().get_pool_size());
+      static_cast<uint32_t>(_total_tcp_count),
+      pool::instance().get_pool_size());
 }
 
 /**
@@ -87,7 +89,8 @@ stream::~stream() noexcept {
   log_v2::tcp()->info(
       "TCP stream destroyed. Still {} configured on a thread pool of {} "
       "threads",
-      _total_tcp_count, pool::instance().get_pool_size());
+      static_cast<uint32_t>(_total_tcp_count),
+      pool::instance().get_pool_size());
   log_v2::tcp()->trace("stream closed");
   if (_connection->socket().is_open())
     _connection->close();
