+++ broker/unified_sql/src/stream_storage.cc
@@ -1,20 +1,20 @@
 /**
-* Copyright 2019-2022 Centreon
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-* For more information : contact@centreon.com
-*/
+ * Copyright 2019-2022 Centreon
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * For more information : contact@centreon.com
+ */
 
 #include <fmt/format.h>
 
@@ -331,7 +331,7 @@ void stream::_unified_sql_process_pb_service_status(
             std::string row;
             if (std::isinf(pd.value()))
               row = fmt::format("({},{},'{}',{})", metric_id, ss.last_check(),
-                                ss.state(),
+                                static_cast<uint32_t>(ss.state()),
                                 pd.value() < 0.0 ? -FLT_MAX : FLT_MAX);
             else if (std::isnan(pd.value()))
               row = fmt::format("({},{},'{}',NULL)", metric_id, ss.last_check(),
